(SPIKE) Enable pluggable JRuby handlers
This commit adds a `get-wrapped-handler` function to the public interface of
the RequestHandlerService protocol. This function can be called on a handler
that:

  - Accepts a Ring request map as its first argument. This request will
    contain interfaces to a JRruby interpreter.

  - Accepts a Ruby request map as its second argument. This map will be
    a transformed version of the Ring request with objects and strings
    cast to a format Ruby can work with easily.

  - Returns a JRubyPuppetResponse instance.

The `get-wrapped-handler` function will wrap the supplied handler in middleware
for borrowing a JRuby and transforming between Ring request maps and the Ruby
response object. This change allows plugins to handle requests using the JRuby
pool in a manner consistent wit the core Puppet Server API.